+++ routes/web/managerarea.php
@@ -3,6 +3,7 @@
 declare(strict_types=1);
 
 use Cortex\Foundation\Http\Controllers\Managerarea\HomeController;
+use Cortex\Foundation\Http\Controllers\Managerarea\GenericController;
 
 Route::domain('{managerarea}')->group(function () {
     Route::name('managerarea.')
@@ -11,5 +12,6 @@ Route::domain('{managerarea}')->group(function () {
 
             // Managerarea Home route
              Route::get('/')->name('home')->uses([HomeController::class, 'index']);
+             Route::post('country')->name('country')->uses([GenericController::class, 'country']);
          });
 });